Change block reward for HF15 & 16 as per LRC-6
This changes the formula for HF15 to 66% service nodes, 24% miner, 10%
foundation, and changes the block reward from the current 28 and a bit
to a fixed 25 LOKI per block.  SN reward and LF reward both increase,
miner reward gets cut substantially.  This change is reflecting that
with checkpoints and blink the security provided by miners is
substantially reduced.

In HF16 (with pulse) we plan to become mining-free, with final rewards
also included here for that fork of 21 LOKI per block with 90% to SNs
and 10% to foundation.  (This is tentative assuming pulse is implemented
in HF16).