Replace ARG NPROC with invoke $(nproc) directly, remove unbound hack
- ARG is a build time variable that you must pass in via docker build
--build-arg NPROC=<number>. What we want is if NPROC is not set, use
$(nproc) otherwise use NPROC passed via command line. You can't do this,
even with setting ENV, because it can't evaluate commands at runtime.

Exporting commands via RUN, only persist for that RUN command as each
RUN sets up a containerized ENV, so you'd have to run an if/else check
for each RUN. There are other hacky work-arounds like writing $(nproc)
to a file and reading it if NPROC isn't passed but all of this is more
trouble than its worth.

When we have an issue with needing to restrict the -j param on make, we
can use one of the workarounds.

- Change openssl to just vanilla make and make install_sw, currently as
is doesn't work in the container, reports make error.