Allow prompting for restore height for hw wallet
For no good reason the hardware wallet creation didn't allow you to
specify a restore height from the cli wallet for no good reason.  This
commit lets it use the same prompt, and improves the prompt somewhat:

- you can now specify `curr` to use the current blockchain height (and
this is the default for the hardware wallet).

- added a note about the purpose of the restore height (i.e. that
transactions won't be seen if before the restore height).

- put example values in the description.

- show the default in the prompt.