Clean up HW debug code
The hard wallet debug code had no way to enable it, and if you did
manually add the define, didn't compile.  It was also nasty, gross,
disgusting code that someone slopped into the file.

This fixes it, adds a cmake option for it, and significantly cleans it
up--just because code is for debugging doesn't mean it should be nasty
and broken.