Remove limit on burned fee
The reason behind the limit is that the burn amount was supposed to be
encoded using varint encoding and therefore the limit was to make sure
that once we figured out the final burn amount and put it in, we were
guaranteed not to be making the TX extra any bigger (just in case that
could end up making the overall tx get a couple bytes bigger and break
the tx size limit).

However, it never actually *used* varint encoding: instead it is encoded
as a raw, full size uint64_t value of 8 bytes regardless of the value,
so this check is not actually doing anything.  (And if we changed it
to a varint we'd break the protocol, so just leave it).

It also turns out that this comment was wrong:

    This value (~4398 OXEN) was chosen because it's unlikely to ever be
    needed to be burned in a single transaction

Also I hear that some users really do need more than 640kB RAM. ;-)