Remove external block generation/submission support
We have no need for any of this, so just remove it rather than convert
it.

(This doesn't drop block generation support, which we can still do via
start/stop mining calls, but rather just the ability to hook up an
external miner, which we are unlikely to ever need again).