Fix moronic code
This `getKeyFromBlockchain` function is completely moronic: it doesn't
do anything with its `reference_index`, and has a commented description
that doesn't match the actual implementation at all.

Moreover the `reference_index` argument, at once place here, is passed a
*function pointer cast to a size_t*.  A generous interpretation is the
author knew that it wasn't being used, and thus could be passed the
abitrary address of a function that happens to return a random value --
but it's all fine because the argument isn't used, right?

Replace it with a .randomize() on ctkey that does the same thing, but
with 97% less moron.

This code is beyond pathetic.  I am starting to see why some people need
Rust to guard their incompetence away from the code they write.