Add 2.6.1 storage server upgrade with 8 day dereg grace period
The storage server update is important enough that we need to force push
it, and so this update bumps the required storage server version to
2.6.1 with the fixes.

Because this is happening less than a week before the fork block, this
also hacks in an 8-day grace period during which 10.4.1 nodes won't
deregister nodes that haven't updated.  (They will still get decommed,
but will just stay there without getting a final dereg until 8 days
after the upgrade).