Remove oxen-encoding, use oxen-mq/external/oxen-encoding
We can avoid keeping 2 copies of the same library since oxen-encoding in oxen-mq
is exposed as a public library allowing oxen-core to interface with
oxen-encoding via oxen-mq.

This simplifies the build and avoids any need to synchronise the
dependencies together (e.g. oxen-encoding was updated to C++20 but
oxen-mq was still on the C++17 branch). We can move up to C++20 in
unison when oxen-mq is updated to C++20 (and hence avoid any
incompatibility issues).