Interim fix `initial_data_set` race by locking in BLS agg
Also since nodes are untrusted this has a pretty big implication on the
current approach. We can't necessarily trust the initial data as the
reference. We more likely have to actually aggregate the responses and
use the majority. This should be fix in a polish-pass of this BLS
integration work.